+++ app.py
@@ -2,7 +2,6 @@ import streamlit as st
 from utils.main_background import main_background
 from main.intro import set_intro
 from main.description import set_description
-from utils.signIn_modal import signIn_modal
 import toml
 
 setting = toml.load('static/toml/setting.toml')
@@ -31,5 +30,4 @@ st.markdown("""
 """, unsafe_allow_html=True)
 
 if st.button("지금 바로 분석 하러가기"):
-    # signIn_modal()
-    st.switch_page("pages/database.py")
+    st.switch_page("pages/chat.py")
